Rebase onto Channel refactoring
