+++ pyiron_workflow/node.py
@@ -9,21 +9,15 @@ from __future__ import annotations
 
 import sys
 import warnings
-from abc import ABC, abstractmethod
+from abc import ABC
 from concurrent.futures import Future
 from importlib import import_module
 from typing import Any, Literal, Optional, TYPE_CHECKING
 
-from pyiron_workflow.channels import (
-    InputSignal,
-    AccumulatingInputSignal,
-    OutputSignal,
-    NOT_DATA,
-)
 from pyiron_workflow.draw import Node as GraphvizNode
 from pyiron_workflow.snippets.files import DirectoryObject
 from pyiron_workflow.has_to_dict import HasToDict
-from pyiron_workflow.io import Signals, IO
+from pyiron_workflow.io import HasIO
 from pyiron_workflow.run import Runnable, ReadinessError
 from pyiron_workflow.semantics import Semantic
 from pyiron_workflow.storage import StorageInterface
@@ -39,12 +33,10 @@ if TYPE_CHECKING:
 
     import graphviz
 
-    from pyiron_workflow.channels import Channel
     from pyiron_workflow.composite import Composite
-    from pyiron_workflow.io import Inputs, Outputs
 
 
-class Node(HasToDict, Semantic, Runnable, ABC, metaclass=AbstractHasPost):
+class Node(HasToDict, Semantic, Runnable, HasIO, ABC, metaclass=AbstractHasPost):
     """
     Nodes are elements of a computational graph.
     They have inputs and outputs to interface with the wider world, and perform some
@@ -296,7 +288,6 @@ class Node(HasToDict, Semantic, Runnable, ABC, metaclass=AbstractHasPost):
             **kwargs: Keyword arguments passed on with `super`.
         """
         super().__init__(*args, label=label, parent=parent, **kwargs)
-        self.signals = self._build_signal_channels()
         self._working_directory = None
         self._storage_backend = None
         self.storage_backend = storage_backend
@@ -335,16 +326,6 @@ class Node(HasToDict, Semantic, Runnable, ABC, metaclass=AbstractHasPost):
                 pass
         # Else neither loading nor running now -- no action required!
         self.graph_root.tidy_working_directory()
-
-    @property
-    @abstractmethod
-    def inputs(self) -> Inputs:
-        pass
-
-    @property
-    @abstractmethod
-    def outputs(self) -> Outputs:
-        pass
 
     @property
     def graph_path(self) -> str:
@@ -582,35 +563,6 @@ class Node(HasToDict, Semantic, Runnable, ABC, metaclass=AbstractHasPost):
         """
         return self.pull(run_parent_trees_too=True, **kwargs)
 
-    def set_input_values(self, **kwargs) -> None:
-        """
-        Match keywords to input channels and update their values.
-
-        Throws a warning if a keyword is provided that cannot be found among the input
-        keys.
-
-        Args:
-            **kwargs: input key - input value (including channels for connection) pairs.
-        """
-        for k, v in kwargs.items():
-            if k in self.inputs.labels:
-                self.inputs[k] = v
-            else:
-                warnings.warn(
-                    f"The keyword '{k}' was not found among input labels. If you are "
-                    f"trying to update a node keyword, please use attribute assignment "
-                    f"directly instead of calling"
-                )
-
-    def _build_signal_channels(self) -> Signals:
-        signals = Signals()
-        signals.input.run = InputSignal("run", self, self.run)
-        signals.input.accumulate_and_run = AccumulatingInputSignal(
-            "accumulate_and_run", self, self.run
-        )
-        signals.output.ran = OutputSignal("ran", self)
-        return signals
-
     @property
     def working_directory(self):
         if self._working_directory is None:
@@ -621,35 +573,9 @@ class Node(HasToDict, Semantic, Runnable, ABC, metaclass=AbstractHasPost):
                 self._working_directory = DirectoryObject(self.label)
         return self._working_directory
 
-    def disconnect(self):
-        """
-        Disconnect all connections belonging to inputs, outputs, and signals channels.
-
-        Returns:
-            [list[tuple[Channel, Channel]]]: A list of the pairs of channels that no
-                longer participate in a connection.
-        """
-        destroyed_connections = []
-        destroyed_connections.extend(self.inputs.disconnect())
-        destroyed_connections.extend(self.outputs.disconnect())
-        destroyed_connections.extend(self.signals.disconnect())
-        return destroyed_connections
-
     @property
     def ready(self) -> bool:
         return super().ready and self.inputs.ready
-
-    @property
-    def connected(self) -> bool:
-        return self.inputs.connected or self.outputs.connected or self.signals.connected
-
-    @property
-    def fully_connected(self):
-        return (
-            self.inputs.fully_connected
-            and self.outputs.fully_connected
-            and self.signals.fully_connected
-        )
 
     @property
     def color(self) -> str:
@@ -724,16 +650,6 @@ class Node(HasToDict, Semantic, Runnable, ABC, metaclass=AbstractHasPost):
             )
         return graph
 
-    def activate_strict_hints(self):
-        """Enable type hint checks for all data IO"""
-        self.inputs.activate_strict_hints()
-        self.outputs.activate_strict_hints()
-
-    def deactivate_strict_hints(self):
-        """Disable type hint checks for all data IO"""
-        self.inputs.deactivate_strict_hints()
-        self.outputs.deactivate_strict_hints()
-
     def __str__(self):
         return (
             f"{self.label} ({self.__class__.__name__}):\n"
@@ -741,163 +657,6 @@ class Node(HasToDict, Semantic, Runnable, ABC, metaclass=AbstractHasPost):
             f"{str(self.outputs)}\n"
             f"{str(self.signals)}"
         )
-
-    def _connect_output_signal(self, signal: OutputSignal):
-        self.signals.input.run.connect(signal)
-
-    def __rshift__(self, other: InputSignal | Node):
-        """
-        Allows users to connect run and ran signals like: `first_node >> second_node`.
-        """
-        other._connect_output_signal(self.signals.output.ran)
-        return other
-
-    def _connect_accumulating_input_signal(self, signal: AccumulatingInputSignal):
-        self.signals.output.ran.connect(signal)
-
-    def __lshift__(self, others):
-        """
-        Connect one or more `ran` signals to `accumulate_and_run` signals like:
-        `this_node << some_node, another_node, or_by_channel.signals.output.ran`
-        """
-        self.signals.input.accumulate_and_run << others
-
-    def copy_io(
-        self,
-        other: Node,
-        connections_fail_hard: bool = True,
-        values_fail_hard: bool = False,
-    ) -> None:
-        """
-        Copies connections and values from another node's IO onto this node's IO.
-        Other channels with no connections are ignored for copying connections, and all
-        data channels without data are ignored for copying data.
-        Otherwise, default behaviour is to throw an exception if any of the other node's
-        connections fail to copy, but failed value copies are simply ignored (e.g.
-        because this node does not have a channel with a commensurate label or the
-        value breaks a type hint).
-        This error throwing/passing behaviour can be controlled with boolean flags.
-
-        In the case that an exception is thrown, all newly formed connections are broken
-        and any new values are reverted to their old state before the exception is
-        raised.
-
-        Args:
-            other (Node): The other node whose IO to copy.
-            connections_fail_hard: Whether to raise exceptions encountered when copying
-                connections. (Default is True.)
-            values_fail_hard (bool): Whether to raise exceptions encountered when
-                copying values. (Default is False.)
-        """
-        new_connections = self._copy_connections(other, fail_hard=connections_fail_hard)
-        try:
-            self._copy_values(other, fail_hard=values_fail_hard)
-        except Exception as e:
-            for this, other in new_connections:
-                this.disconnect(other)
-            raise e
-
-    def _copy_connections(
-        self,
-        other: Node,
-        fail_hard: bool = True,
-    ) -> list[tuple[Channel, Channel]]:
-        """
-        Copies all the connections in another node to this one.
-        Expects all connected channels on the other node to have a counterpart on this
-        node -- i.e. the same label, type, and (for data) a type hint compatible with
-        all the existing connections being copied.
-        This requirement can be optionally relaxed such that any failures encountered
-        when attempting to make a connection (i.e. this node has no channel with a
-        corresponding label as the other node, or the new connection fails its validity
-        check), such that we simply continue past these errors and make as many
-        connections as we can while ignoring errors.
-
-        This node may freely have additional channels not present in the other node.
-        The other node may have additional channels not present here as long as they are
-        not connected.
-
-        If an exception is going to be raised, any connections copied so far are
-        disconnected first.
-
-        Args:
-            other (Node): the node whose connections should be copied.
-            fail_hard (bool): Whether to raise an error an exception is encountered
-                when trying to reproduce a connection. (Default is True; revert new
-                connections then raise the exception.)
-
-        Returns:
-            list[tuple[Channel, Channel]]: A list of all the newly created connection
-                pairs (for reverting changes).
-        """
-        new_connections = []
-        for my_panel, other_panel in [
-            (self.inputs, other.inputs),
-            (self.outputs, other.outputs),
-            (self.signals.input, other.signals.input),
-            (self.signals.output, other.signals.output),
-        ]:
-            for key, channel in other_panel.items():
-                for target in channel.connections:
-                    try:
-                        my_panel[key].connect(target)
-                        new_connections.append((my_panel[key], target))
-                    except Exception as e:
-                        if fail_hard:
-                            # If you run into trouble, unwind what you've done
-                            for this, other in new_connections:
-                                this.disconnect(other)
-                            raise e
-                        else:
-                            continue
-        return new_connections
-
-    def _copy_values(
-        self,
-        other: Node,
-        fail_hard: bool = False,
-    ) -> list[tuple[Channel, Any]]:
-        """
-        Copies all data from input and output channels in the other node onto this one.
-        Ignores other channels that hold non-data.
-        Failures to find a corresponding channel on this node (matching label, type, and
-        compatible type hint) are ignored by default, but can optionally be made to
-        raise an exception.
-
-        If an exception is going to be raised, any values updated so far are
-        reverted first.
-
-        Args:
-            other (Node): the node whose data values should be copied.
-            fail_hard (bool): Whether to raise an error an exception is encountered
-                when trying to duplicate a value. (Default is False, just keep going
-                past other's channels with no compatible label here and past values
-                that don't match type hints here.)
-
-        Returns:
-            list[tuple[Channel, Any]]: A list of tuples giving channels whose value has
-                been updated and what it used to be (for reverting changes).
-        """
-        old_values = []
-        for my_panel, other_panel in [
-            (self.inputs, other.inputs),
-            (self.outputs, other.outputs),
-        ]:
-            for key, to_copy in other_panel.items():
-                if to_copy.value is not NOT_DATA:
-                    try:
-                        old_value = my_panel[key].value
-                        my_panel[key].value = to_copy.value  # Gets hint-checked
-                        old_values.append((my_panel[key], old_value))
-                    except Exception as e:
-                        if fail_hard:
-                            # If you run into trouble, unwind what you've done
-                            for channel, value in old_values:
-                                channel.value = value
-                            raise e
-                        else:
-                            continue
-        return old_values
 
     def replace_with(self, other: Node | type[Node]):
         """
@@ -916,23 +675,6 @@ class Node(HasToDict, Semantic, Runnable, ABC, metaclass=AbstractHasPost):
             self.parent.replace_child(self, other)
         else:
             warnings.warn(f"Could not replace_child {self.label}, as it has no parent.")
-
-    def __setstate__(self, state):
-        super().__setstate__(state)
-
-        # Channels don't store their own node in their state, so repopulate it
-        for io_panel in self._owned_io_panels:
-            for channel in io_panel:
-                channel.owner = self
-
-    @property
-    def _owned_io_panels(self) -> list[IO]:
-        return [
-            self.inputs,
-            self.outputs,
-            self.signals.input,
-            self.signals.output,
-        ]
 
     @property
     def class_name(self) -> str:
